+++ jubatus/server/server/regression_types.hpp
@@ -19,11 +19,11 @@ namespace jubatus {
 struct scored_datum {
  public:
   MSGPACK_DEFINE(score, data);
-  float score;
+  double score;
   jubatus::core::fv_converter::datum data;
   scored_datum() {
   }
-  scored_datum(float score, const jubatus::core::fv_converter::datum& data)
+  scored_datum(double score, const jubatus::core::fv_converter::datum& data)
     : score(score), data(data) {
   }
 };
